WIP. Split out Favicon component
